# Improve `create_submesh` performance