The current implementation of `create_submesh` does not scale due to an `std::find` on a potentially large array in a tight loop. This PR precomputes the map to avoid this issue.

Performance results:

I ran the following code for different values of `n`:
```
msh = mesh.create_unit_cube(
        MPI.COMM_WORLD, n, n, n, ghost_mode=mesh.GhostMode.none)
tdim = msh.topology.dim
entities = mesh.locate_entities(msh, tdim, lambda x: x[0] <= 0.5)
mesh.create_submesh(msh, tdim, entities)
```
The timings are as follows:

|  | `n = 8` | `n = 16` | `n = 32` |
| ------------- | ------------- | ------------- | ------------- |
| Main  | 0.0365 s  | 0.861 s  | 46.6 s |
| This branch  | 0.00904 s  | 0.0459 s  | 0.38 s  |

The difference in performance gets larger as the mesh size increases.